# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CareConnect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: getClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /caregivers:
    get:
      operationId: getCaregivers
      tags: [caregivers]
      summary: List all caregivers
      responses:
        "200":
          description: List of caregivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Caregiver"
    post:
      operationId: createCaregiver
      tags: [caregivers]
      summary: Create a caregiver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaregiverBody"
      responses:
        "201":
          description: Caregiver created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
  /caregivers/{id}:
    get:
      operationId: getCaregiver
      tags: [caregivers]
      summary: Get a caregiver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Caregiver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
        "404":
          description: Not found
    patch:
      operationId: updateCaregiver
      tags: [caregivers]
      summary: Update a caregiver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaregiverBody"
      responses:
        "200":
          description: Updated caregiver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Caregiver"
    delete:
      operationId: deleteCaregiver
      tags: [caregivers]
      summary: Delete a caregiver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /appointments:
    get:
      operationId: getAppointments
      tags: [appointments]
      summary: List all appointments
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /call-logs:
    get:
      operationId: getCallLogs
      tags: [call-logs]
      summary: List all call logs
      responses:
        "200":
          description: List of call logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CallLog"
    post:
      operationId: createCallLog
      tags: [call-logs]
      summary: Create a call log record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallLogBody"
      responses:
        "201":
          description: Call log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
  /call-logs/{id}:
    get:
      operationId: getCallLog
      tags: [call-logs]
      summary: Get a call log by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallLog"
        "404":
          description: Not found
...
  /escalations:
    get:
      operationId: getEscalations
      tags: [escalations]
      summary: List all escalations
      responses:
        "200":
          description: List of escalations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Escalation"
    post:
      operationId: createEscalation
      tags: [escalations]
      summary: Create an escalation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEscalationBody"
      responses:
        "201":
          description: Escalation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
  /escalations/{id}:
    get:
      operationId: getEscalation
      tags: [escalations]
      summary: Get an escalation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Escalation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
        "404":
          description: Not found
    patch:
      operationId: updateEscalation
      tags: [escalations]
      summary: Update an escalation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEscalationBody"
      responses:
        "200":
          description: Updated escalation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Escalation"
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: string
        phone:
          type: string
        preferredLanguage:
          type: string
        carePlan:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_hold]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - address
        - phone
        - preferredLanguage
        - status
        - createdAt
        - updatedAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: string
        phone:
          type: string
        preferredLanguage:
          type: string
        carePlan:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_hold]
        notes:
          type: ["string", "null"]
      required:
        - name
        - address
        - phone
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        dateOfBirth:
          type: ["string", "null"]
        address:
          type: string
        phone:
          type: string
        preferredLanguage:
          type: string
        carePlan:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_hold]
        notes:
          type: ["string", "null"]
    Caregiver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [caregiver, nurse, coordinator, supervisor]
        availability:
          type: string
          enum: [full_time, part_time, on_call]
        languages:
          type: string
        certifications:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave]
        hireDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - phone
        - role
        - availability
        - languages
        - status
        - createdAt
        - updatedAt
    CreateCaregiverBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [caregiver, nurse, coordinator, supervisor]
        availability:
          type: string
          enum: [full_time, part_time, on_call]
        languages:
          type: string
        certifications:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave]
        hireDate:
          type: ["string", "null"]
      required:
        - name
        - phone
        - role
        - availability
    UpdateCaregiverBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [caregiver, nurse, coordinator, supervisor]
        availability:
          type: string
          enum: [full_time, part_time, on_call]
        languages:
          type: string
        certifications:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, on_leave]
        hireDate:
          type: ["string", "null"]
    Appointment:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        caregiverId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [daily_assistance, companionship, transportation, telehealth, medication_management, personal_care]
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
        clientName:
          type: ["string", "null"]
        caregiverName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientId
        - scheduledAt
        - durationMinutes
        - type
        - status
        - createdAt
        - updatedAt
    CreateAppointmentBody:
      type: object
      properties:
        clientId:
          type: integer
        caregiverId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [daily_assistance, companionship, transportation, telehealth, medication_management, personal_care]
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
      required:
        - clientId
        - scheduledAt
        - durationMinutes
        - type
    UpdateAppointmentBody:
      type: object
      properties:
        clientId:
          type: integer
        caregiverId:
          type: ["integer", "null"]
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        type:
          type: string
          enum: [daily_assistance, companionship, transportation, telehealth, medication_management, personal_care]
        status:
          type: string
          enum: [scheduled, confirmed, in_progress, completed, cancelled, no_show]
        notes:
          type: ["string", "null"]
    CallLog:
      type: object
      properties:
        id:
          type: integer
        callerPhone:
          type: ["string", "null"]
        callerName:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        intent:
          type: ["string", "null"]
          enum: [service_inquiry, appointment_request, medication_reminder, urgent_issue, general_inquiry, family_update, null]
        language:
          type: string
        durationSeconds:
          type: ["integer", "null"]
        outcome:
          type: string
          enum: [resolved, escalated, voicemail, transferred, follow_up_needed]
        vapiCallId:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - language
        - outcome
        - createdAt
        - updatedAt
    CreateCallLogBody:
      type: object
      properties:
        callerPhone:
          type: ["string", "null"]
        callerName:
          type: ["string", "null"]
        transcript:
          type: ["string", "null"]
        intent:
          type: ["string", "null"]
        language:
          type: string
        durationSeconds:
          type: ["integer", "null"]
        outcome:
          type: string
          enum: [resolved, escalated, voicemail, transferred, follow_up_needed]
        vapiCallId:
          type: ["string", "null"]
        clientId:
          type: ["integer", "null"]
      required:
        - outcome
...
    Escalation:
      type: object
      properties:
        id:
          type: integer
        callLogId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        assignedTo:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        reason:
          type: string
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - priority
        - reason
        - resolved
        - createdAt
        - updatedAt
    CreateEscalationBody:
      type: object
      properties:
        callLogId:
          type: ["integer", "null"]
        clientId:
          type: ["integer", "null"]
        assignedTo:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        reason:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - priority
        - reason
    UpdateEscalationBody:
      type: object
      properties:
        assignedTo:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, critical]
        reason:
          type: string
        resolved:
          type: boolean
        resolvedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...